# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface WhoisResult {
  /** @nullable */
  domainName: string | null;
  /** @nullable */
  registrar: string | null;
  /** @nullable */
  registrarUrl: string | null;
  /** @nullable */
  createdDate: string | null;
  /** @nullable */
  updatedDate: string | null;
  /** @nullable */
  expiresDate: string | null;
  status: string[];
  nameservers: string[];
  /** @nullable */
  registrantCountry: string | null;
  /** @nullable */
  domainAge: number | null;
  /** @nullable */
  rawText: string | null;
}

export interface DnsRecord {
  type: string;
  value: string;
  /** @nullable */
  ttl?: number | null;
  /** @nullable */
  priority?: number | null;
}

export interface DnsResult {
  domain: string;
  A: DnsRecord[];
  AAAA: DnsRecord[];
  MX: DnsRecord[];
  TXT: DnsRecord[];
  NS: DnsRecord[];
  CNAME: DnsRecord[];
}

export interface TldAvailability {
  tld: string;
  available: boolean;
  domain: string;
}

export interface DomainAvailabilityResult {
  baseDomain: string;
  results: TldAvailability[];
}

export interface SearchHistory {
  id: number;
  domain: string;
  searchedAt: string;
  /** @nullable */
  ipAddress: string | null;
}

export interface PopularDomain {
  domain: string;
  count: number;
}

export type GetSearchHistoryParams = {
limit?: number;
};