# lib/api-zod/src/generated/types/dnsRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DnsRecord {
  type: string;
  value: string;
  /** @nullable */
  ttl?: number | null;
  /** @nullable */
  priority?: number | null;
}