# lib/api-zod/src/generated/types/dnsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DnsRecord } from './dnsRecord';

export interface DnsResult {
  domain: string;
  A: DnsRecord[];
  AAAA: DnsRecord[];
  MX: DnsRecord[];
  TXT: DnsRecord[];
  NS: DnsRecord[];
  CNAME: DnsRecord[];
}