# lib/api-zod/src/generated/types/domainAvailabilityResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TldAvailability } from './tldAvailability';

export interface DomainAvailabilityResult {
  baseDomain: string;
  results: TldAvailability[];
}