# lib/api-zod/src/generated/types/searchHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchHistory {
  id: number;
  domain: string;
  searchedAt: string;
  /** @nullable */
  ipAddress: string | null;
}