# lib/api-zod/src/generated/types/tldAvailability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TldAvailability {
  tld: string;
  available: boolean;
  domain: string;
}