# lib/api-zod/src/generated/types/whoisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WHOIS Lookup API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhoisResult {
  /** @nullable */
  domainName: string | null;
  /** @nullable */
  registrar: string | null;
  /** @nullable */
  registrarUrl: string | null;
  /** @nullable */
  createdDate: string | null;
  /** @nullable */
  updatedDate: string | null;
  /** @nullable */
  expiresDate: string | null;
  status: string[];
  nameservers: string[];
  /** @nullable */
  registrantCountry: string | null;
  /** @nullable */
  domainAge: number | null;
  /** @nullable */
  rawText: string | null;
}